Refactoring Models into interface directory
