+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "device-monitor",
-  "version": "0.0.",
+  "version": "0.0.2",
   "private": true,
   "dependencies": {
     "@capacitor/android": "^2.1.2",
@@ -20,9 +20,12 @@
     "ionicons": "^5.0.0",
     "react": "^16.13.0",
     "react-dom": "^16.13.0",
+    "react-redux": "^7.2.0",
     "react-router": "^5.1.2",
     "react-router-dom": "^5.1.2",
     "react-scripts": "3.4.1",
+    "redux": "^4.0.5",
+    "redux-thunk": "^2.3.0",
     "typescript": "3.8.3"
   },
   "scripts": {
