+++ src/drivers/interface/index.tsx
@@ -1,6 +1,6 @@
 /* istanbul ignore file */
 
-import {DeviceDetailsModel} from '../../models/';
+import {DeviceDetailsModel} from '../../interfaces/models';
 
 export interface IDeviceBridge {
     getDeviceInfo(): Promise<DeviceDetailsModel>;
