+++ src/features/device-summary/DeviceDetails.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import DeviceDetails from './DeviceDetails';
-import {DeviceDetailsModel} from '../../models';
+import {DeviceDetailsModel} from '../../interfaces/models';
 
 describe('DeviceField', () => {
     test('rendering a simple device field', async () => {
