+++ src/store/actions/device-details-action.ts
@@ -1,0 +1,7 @@
+
+import { DeviceDetailsModel } from '../../interfaces/models';
+
+/*export const fetchDeviceDetailsStartAction = createAction('fetch inventory list :: start');
+export const fetchDevicesDetailsSuccessAction = createAction<IDeviceFlatModel[]>('fetch inventory list :: success');
+export const fetchDevicesDetailsErrorAction = createAction<string[]>('fetch inventory list :: error');
+*/
